% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDiscreteField.R
\name{ImportDiscreteField}
\alias{ImportDiscreteField}
\title{Generate a data frame with a station's discrete field data.}
\usage{
ImportDiscreteField(stationNumber)
}
\arguments{
\item{stationNumber}{A station's FLIMS* station number.}
}
\value{
A data frame with all of the discrete field data records for a station.
}
\description{
Generate a data frame with a station's discrete field data.
}
\examples{
test_discrete_field <- "B9D82851352"
ImportDiscreteField(test_discrete_field)
}
