% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportDiscreteLab.R
\name{ImportDiscreteLab}
\alias{ImportDiscreteLab}
\title{Generate a data frame with a station's discrete lab data.}
\usage{
ImportDiscreteLab(stationNumber)
}
\arguments{
\item{stationNumber}{a station's FLIMS* station number.}
}
\value{
A data frame with all of the discrete lab data records for a station.
}
\description{
Generate a data frame with a station's discrete lab data.
}
\examples{
test_discrete_lab <- "B9D82851352"
ImportDiscreteLab(test_discrete_lab)
}
