% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportContinuousCSV.R
\name{ImportContinuousCSV}
\alias{ImportContinuousCSV}
\title{Import CSV links from CNRA DWR Continuous Data Download Links}
\usage{
ImportContinuousCSV(csv_link)
}
\arguments{
\item{csv_link}{A CSV link sourced from the CNRA Station-Trace links to time-series downloads.}
}
\value{
A formatted data frame which excludes quality code column and other extraneous information.
}
\description{
Input a CSV link from CNRA and return a formatted data frame for that file.
}
\examples{
CNRA_CSV_link <- "https://wdlstorageaccount.blob.core.windows.net/continuous/01N04E36Q001M/por/01N04E36Q001M_Groundwater_Level_Below_Ground_Surface_Raw.csv"
ImportContinuousCSV(csv_link = CNRA_CSV_link)
}
