% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportContinuousMean.R
\name{ImportContinuousMean}
\alias{ImportContinuousMean}
\title{Generate a data frame with a station's continuous daily average data.}
\usage{
ImportContinuousMean(stationNumber, parameters = NULL)
}
\arguments{
\item{stationNumber}{A station's station number.}

\item{parameters}{A character vector of the parameters you want to select.}
}
\value{
A data frame with all or some of the parameters for a station's daily average data.
}
\description{
Generate a data frame with a station's continuous daily average data.
}
\examples{
test_station_1 <- "B9156000"
test_station_2 <- "B91560Q"
ImportContinuousMean(stationNumber = test_station_1)
ImportContinuousMean(stationNumber = test_station_2)
test_parameters <- c("Water Temperature",
                     "Electrical Conductivity at 25C",
                     "Dissolved Oxygen")
ImportContinuousMean(stationNumber = test_station_1, parameters = test_parameters)
ImportContinuousMean(stationNumber = test_station_2, parameters = test_parameters)
}
