% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportContinuousRaw.R
\name{ImportContinuousRaw}
\alias{ImportContinuousRaw}
\title{Generate a data frame with a station's continuous raw data.}
\usage{
ImportContinuousRaw(stationNumber, parameters = NULL)
}
\arguments{
\item{stationNumber}{A station's station number.}

\item{parameters}{A character vector of the parameters you want to select.}
}
\value{
A data frame with all or some of the parameters for a station's raw data.
}
\description{
Input a station number* to import a formatted data frame with its continuous raw data from WDL.
}
\examples{
test_station_1 <- "B9156000"
ImportContinuousRaw(stationNumber = test_station_1)
test_parameters <- c("Water Temperature",
                     "Electrical Conductivity at 25C",
                     "Dissolved Oxygen")
ImportContinuousRaw(stationNumber = test_station_1, parameters = test_parameters)
}
